final validation, beginning of cms creation

+++ backend/app/Http/Controllers/AuthController.php
@@ -3,7 +3,6 @@
 namespace App\Http\Controllers;
 
 use App\Models\User;
-use Hash;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -11,18 +10,26 @@ class AuthController extends Controller
 {
     public function register (Request $request){
         $validate = $request->validate([
-            'name' => 'required|string|max:255',
+            'name' => 'required|string|max:255|min:5',
             'email' => 'required|string|email|max:255',
             'password' => 'required|string|max:255|min:7',
         ]);
 
-        $user = User::create([
-            'name' => $validate['name'],
-            'email' => $validate['email'],
-            'password' => Hash::make($validate['password'])
-        ]);
-        
-        return response()->json(['message'=>'User registered successfully!'],201);
+        try {
+                $user = User::create([
+                    'name' => $validate['name'],
+                    'email' => $validate['email'],
+                    'password' => $validate['password'], 
+                ]);
+                return response()->json(['message' => 'User registered successfully!'], 201);
+    
+            } 
+            catch (\Illuminate\Database\QueryException $e) {
+                if ($e->getCode() === '23000') {
+                    return response()->json(['message' => 'Email already exists. Please use a different email address.'], 409);
+                }
+            }
+            return response()->json(['error' => 'An error occurred while registering. Please try again later.'], 500);
 
     }
 
@@ -31,7 +38,7 @@ class AuthController extends Controller
         $credentials = $request->validate([
             'email' => 'required|email|string|max:255',
             'password' => 'required',
-        ]);
+        ]); 
 
         if(!Auth::attempt($credentials)){
             return response()->json(['message'=> 'Invalid credentials!'],401);
@@ -44,7 +51,10 @@ class AuthController extends Controller
     }
 
     public function logout(Request $request){
-        $request->user()->tokens()->delete();
+        $user = auth()->user();
+        if($user){
+            $request->user()->tokens()->delete();
+        }
         return response()->json(['message' => 'Logged out successfully!'],200);
     }
 }
